Modify Battery Percentage Type

+++ sender/battery/src/main.cpp
@@ -11,7 +11,7 @@ int main() {
     
     while (1) {
         std::cout<<"voltage : "<<piracer.getVoltage()<<std::endl;
-        battery_percent = ((((piracer.getVoltage() / 3) - 3.1) / 1.1) * 100);
+        battery_percent = (uint8_t)((((piracer.getVoltage() / 3) - 3.1) / 1.1) * 100);
         std::cout<<"percent : "<<battery_percent<<std::endl;
         battery.setBattery(battery_percent);
         usleep(10000);
